requirements.txt
tgcrypto
asyncio
ffmpeg-python
googletrans==4.0.0rc1
pytgcalls==3.0.0.dev16
youtube_dl
pafy==0.5.5
pytube==11.0.1
git+https://github.com/pyrogram/pyrogram@master
youtube_search_python==1.4.6
youtube_search
python-dotenv
aiofiles
aiohttp
requests
psutil
future
gTTS
wget
